updating path files within the relocated files
