+++ client/src/components/AddProjectPage/AddProject.js
@@ -5,7 +5,7 @@ import {Form,Button} from 'react-bootstrap'
 import './Style.scss';
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import InputTag from './InputTag'
+import InputTag from '../InputTag'
 import axios from 'axios'
 
 class AddProject extends React.Component {
