+++ client/src/components/AddProjectPage/CreateProjectContainer.js
@@ -1,7 +1,7 @@
 // CreateProject.js
 
 import { connect } from 'react-redux';
-import { createProject } from '../actions';
+import { createProject } from '../../actions';
 import AddProject from './AddProject';
 
 const mapDispatchToProps = dispatch => {
